excluding endpoints with no file to download from the checks

+++ checkDownloads.js
@@ -15,6 +15,7 @@ const iterateChecker = async () => {
     console.log(errorLogColour, "No files downloaded")
     return
   }
+  console.log("Checking Downloads")
 
   const yearValue = {
     5: "2018-19",
@@ -24,6 +25,9 @@ const iterateChecker = async () => {
     9: "2022-23",
   }
 
+  let filesChecked = 0
+  let filesExcluded = 0
+
   for (const state of states) {
     for (const district of state.districts) {
       for (const block of district.blocks) {
@@ -31,47 +35,70 @@ const iterateChecker = async () => {
           for (let i = 5; i <= 9; i++) {
             const currentYear = i
 
-            const currentFileName = path.join(
-              downloadsDir,
-              `${school.districtId}-${school.blockId}_${
-                yearValue[currentYear]
-              }_${school.schoolName.replace(/[/?<>\\:*|"\s]/g, "-")}`,
-            )
+            // ignore endpoints where no file is expected for download
+            if (
+              school[
+                `isOperational${yearValue[currentYear].replace("-", "")}`
+              ] === 0
+            ) {
+              const currentFileName = path.join(
+                downloadsDir,
+                `${school.districtId}-${school.blockId}_${
+                  yearValue[currentYear]
+                }_${school.schoolName.replace(/[/?<>\\:*|"\s]/g, "-")}`,
+              )
 
-            const newFileName = path.join(
-              downloadsDir,
-              `${school.schoolId}_${
-                yearValue[currentYear]
-              }_${school.schoolName.replace(/[/?<>\\:*|"\s]/g, "-")}`,
-            )
+              const newFileName = path.join(
+                downloadsDir,
+                `${school.schoolId}_${
+                  yearValue[currentYear]
+                }_${school.schoolName.replace(/[/?<>\\:*|"\s]/g, "-")}`,
+              )
 
-            if (fs.existsSync(currentFileName)) {
-              const content = fs.readFileSync(currentFileName)
-              if (content.length === 0) {
-                await new Promise(async resolve => {
-                  fs.unlinkSync(currentFileName)
-                  console.groupCollapsed(path.basename(currentFileName))
-                  const trigger = await schoolDownload(
-                    school,
-                    currentYear,
-                    currentFileName,
-                  )
+              if (fs.existsSync(currentFileName)) {
+                const content = fs.readFileSync(currentFileName)
+
+                // retry downloading if the file is corrupted
+                if (content.length === 0) {
+                  console.groupCollapsed("Retrying")
+
+                  await new Promise(async resolve => {
+                    fs.unlinkSync(currentFileName)
+                    const trigger = await schoolDownload(
+                      school,
+                      currentYear,
+                      currentFileName,
+                    )
+                    resolve(trigger)
+                  })
+
                   console.groupEnd()
-                  resolve(trigger)
-                })
-              }
+                }
 
-              fs.renameSync(currentFileName, newFileName)
+                filesChecked++
+                // rename the file after checking
+                fs.renameSync(currentFileName, newFileName)
+              } else {
+                filesChecked++
+                // retry downloading if file is missing
+                console.groupCollapsed(newFileName)
+                schoolDownload(school, currentYear, newFileName)
+                console.groupEnd()
+              }
             } else {
-              console.groupCollapsed(newFileName)
-              schoolDownload(school, currentYear, newFileName)
-              console.groupEnd()
+              filesExcluded++
             }
           }
         }
       }
     }
   }
+
+  console.log(
+    `${filesChecked} files checked + ${filesExcluded} endpoints without an available file = ${
+      filesChecked + filesExcluded
+    } entries processed`,
+  )
 }
 
 iterateChecker()
